+++ src/leoIntegration.ts
@@ -55,6 +55,7 @@ export class LeoIntegration {
 
     // * State flags
     public leoStates: LeoStates;
+    public isExecuteScript = false; // An execute script (CTRL+B) is running
 
     // * UNL link provider
     public linkProvider: UnlProvider;
@@ -663,9 +664,9 @@ export class LeoIntegration {
             if (this.config.startServerAutomatically) {
                 if (this.config.limitUsers > 1) {
                     utils.findSingleAvailablePort(this.config.connectionPort)
-                        .then((p_availablePort) => {
+                        .then((_p_availablePort) => {
                             this.startServer();
-                        }, (p_reason) => {
+                        }, (_p_reason) => {
                             // Rejected: Multi user port IN USE so skip start
                             if (this.config.connectToServerAutomatically) {
                                 // Still try to connect if auto-connect is 'on'
@@ -686,7 +687,7 @@ export class LeoIntegration {
 
     public async findInstallPath(): Promise<string> {
 
-        return new Promise((p_resolve, p_reject) => {
+        return new Promise((p_resolve, _p_reject) => {
             // Find Leo's installation folder with command line tricks from Matt Wilkie 
             const options = { encoding: 'utf-8' as BufferEncoding };
             const isWindows = process.platform === 'win32';
@@ -777,7 +778,7 @@ export class LeoIntegration {
                 this.config.connectionPort
             )
             .then(
-                (p_message) => {
+                (_p_message) => {
                     utils.setContext(Constants.CONTEXT_FLAGS.SERVER_STARTED, true); // server started
                     if (this.config.connectToServerAutomatically) {
                         setTimeout(() => {
@@ -869,7 +870,7 @@ export class LeoIntegration {
                             // utils.setContext(Constants.CONTEXT_FLAGS.AUTO_CONNECT, true);
                             setTimeout(() => {
                                 // Try to open last opened files, if any
-                                this._openLastFiles().catch(error => {
+                                this._openLastFiles().catch(_error => {
                                     // No open files
                                 });
                             }, 0);
@@ -1121,7 +1122,7 @@ export class LeoIntegration {
                 return this.sendAction(
                     Constants.LEOBRIDGE.HANDLE_UNL,
                     { unl: p_arg.unl }
-                ).then((p_result: LeoBridgePackage) => {
+                ).then((_p_result: LeoBridgePackage) => {
                     this.setupRefresh(
                         Focus.NoChange,
                         {
@@ -1145,6 +1146,185 @@ export class LeoIntegration {
         catch (e) {
             console.log('FAILED HANDLE URL! ', p_arg);
         }
+    }
+
+    public async gotoLineInLeoOutline(p_arg: any): Promise<any> {
+        // When the active editor is an external file referenced by an @file node 
+        // in the current Leo outline, this method finds that @file node,
+        // selects it in the Leo outline, and attempts to place the cursor
+        // to the corresponding line number in the body pane of the selected node.
+        // It processes the first matching @file node found.
+
+        if (!utils.compareVersions(this.serverVersion, { major: 1, minor: 0, patch: 12 })) {
+            return Promise.resolve();
+        }
+
+        const editor = vscode.window.activeTextEditor;
+        if (!editor) {
+            void vscode.window.showErrorMessage('No active editor');
+            return;
+        }
+
+        const document = editor.document;
+        const position = editor.selection.active;
+        const lineNumber = position.line; // 0-indexed
+        let filePath = utils.capitalizeDrive(document.uri.fsPath);
+
+        await this.triggerBodySave(true);
+        try {
+
+            const p_result: LeoBridgePackage = await this.sendAction(
+                Constants.LEOBRIDGE.GOTO_LINE_IN_LEO_OUTLINE,
+                { filePath, lineNumber }
+            );
+
+            if (p_result['found']) {
+                await this.showBody(false);
+                this.setupRefresh(
+                    Focus.Body,
+                    {
+                        tree: true,
+                        body: true,
+                        states: true,
+                        scroll: true, // Scroll to the position in the body
+                        buttons: false,
+                        documents: false,
+                        goto: false,
+                    },
+                    p_result.node || this.lastSelectedNode
+                );
+                this.launchRefresh();
+            } else {
+                // Not found. Offer to import as an @auto, an @clean, or to cancel.
+                // Open modal dialog with options.
+                const choices = [
+                    { title: '@clean', },
+                    { title: '@edit', },
+                    { title: '@asis', },
+                    { title: 'Cancel', isCloseAffordance: true },
+                ];
+                const selection = await vscode.window.showInformationMessage( // Added await
+                    `The file "${filePath}" was not found in the current Leo outline.`,
+                    {
+                        detail: 'Import with...',
+                        modal: true
+                    },
+                    ...choices
+                );
+                if (selection?.title.startsWith('@')) {
+
+                    // get which import type was selected
+                    const importType = selection.title; // '@clean', '@edit', or '@asis'
+
+                    const insertResult = await this.sendAction(
+                        Constants.LEOBRIDGE.INSERT_FILE_NODE,
+                        { filePath, importType }
+                    );
+
+                    const refreshResult = await this.sendAction(
+                        Constants.LEOBRIDGE.REFRESH_FROM_DISK_PNODE,
+                        utils.buildNodeCommand(insertResult.node || this.lastSelectedNode!)
+                    );
+
+                    if (insertResult && refreshResult) {
+
+                        // After import, show the line in Leo outline again
+                        const secondShowLine = await this.sendAction(
+                            Constants.LEOBRIDGE.GOTO_LINE_IN_LEO_OUTLINE,
+                            { filePath, lineNumber }
+                        );
+
+                        if (!secondShowLine['found']) {
+                            vscode.window.showInformationMessage(
+                                `The file "${filePath}" was not found in the current Leo outline after import.`
+                            );
+                        } else {
+                            await this.showBody(false);
+                        }
+                        // refresh the outline and body
+                        this.setupRefresh(
+                            Focus.Body,
+                            {
+                                tree: true,
+                                body: true,
+                                goto: true,
+                                scroll: !!secondShowLine['found'], // Scroll to the position in the body if found
+                                states: true,
+                                documents: true,
+                                buttons: true
+                            },
+                            secondShowLine.node || this.lastSelectedNode
+                        );
+                        return this.launchRefresh();
+                    }
+
+                } else {
+                    // Handle Cancel or dismiss
+                    console.log('User cancelled or dismissed the dialog');
+                }
+            }
+        }
+        catch (e) {
+            console.log('FAILED SHOW LINE IN LEO OUTLINE', p_arg);
+        }
+    }
+
+    public async importIntoLeoOutline(p_arg: vscode.Uri): Promise<any> {
+
+        let filePath = utils.capitalizeDrive(p_arg.fsPath);
+        const choices = [
+            'Default Import',
+            'As @auto',
+            'As @clean',
+            'As @edit',
+            'As @asis',
+        ];
+
+        let result: LeoBridgePackage;
+
+        // * Use quick pick to ask for import type
+        const selection = await vscode.window.showQuickPick(choices, {
+            placeHolder: 'Import into Leo Outline',
+        });
+
+        if (selection && selection === "Default Import") {
+            result = await this.sendAction(
+                Constants.LEOBRIDGE.IMPORT_ANY_FILE,
+                { filenames: [filePath] }
+            );
+
+        } else if (selection && selection.startsWith("As ")) {
+
+            const importType = selection.split(' ')[1].toLowerCase(); // '@auto', '@clean', '@edit', or '@asis'
+
+            const insertResult = await this.sendAction(
+                Constants.LEOBRIDGE.INSERT_FILE_NODE,
+                { filePath, importType }
+            );
+
+            result = await this.sendAction(
+                Constants.LEOBRIDGE.REFRESH_FROM_DISK_PNODE,
+                utils.buildNodeCommand(insertResult.node || this.lastSelectedNode!)
+            );
+        } else {
+            // Handle Cancel or dismiss
+            return Promise.resolve();
+        }
+
+        this.setupRefresh(
+            Focus.Outline, // No change in focus, but refresh all panels
+            {
+                tree: true,
+                body: true,
+                states: true,
+                buttons: false,
+                documents: true,
+                goto: false,
+            },
+            result.node || this.lastSelectedNode
+        );
+        return this.launchRefresh();
+
     }
 
     /**
@@ -1789,7 +1969,7 @@ export class LeoIntegration {
      * React to the closing of 'tabs' via
      * events from vscode.window.tabGroups
      */
-    public _onTabGroupsChanged(p_event: vscode.TabGroupChangeEvent): void {
+    public _onTabGroupsChanged(_p_event: vscode.TabGroupChangeEvent): void {
         this.checkClosedTabs(); // debounced
     }
 
@@ -1797,7 +1977,7 @@ export class LeoIntegration {
      * React to the closing of 'tabgroups' via
      * events from vscode.window.tabGroups
      */
-    public _onTabsChanged(p_event: vscode.TabChangeEvent): void {
+    public _onTabsChanged(_p_event: vscode.TabChangeEvent): void {
         this.checkClosedTabs(); // debounced
     }
 
@@ -1858,9 +2038,9 @@ export class LeoIntegration {
 
     /**
      * * Whole window has been minimized/restored
-     * @param p_windowState the state of the window that changed
+     * @param _p_windowState the state of the window that changed
      */
-    public _changedWindowState(p_windowState: vscode.WindowState): void {
+    public _changedWindowState(_p_windowState: vscode.WindowState): void {
         // no other action
         this.triggerBodySave(true, true);
     }
@@ -2356,7 +2536,7 @@ export class LeoIntegration {
             if (this._onDidHideResolve) {
                 console.error('IN triggerBodySave AND _onDidHideResolve PROMISE ALREADY EXISTS!');
             }
-            const w_resolveAfterEditHeadline = new Promise<void>((p_resolve, p_reject) => {
+            const w_resolveAfterEditHeadline = new Promise<void>((p_resolve, _p_reject) => {
                 this._onDidHideResolve = p_resolve;
             });
             return w_resolveAfterEditHeadline;
@@ -3194,7 +3374,7 @@ export class LeoIntegration {
                         focus: w_focusTree
                     }).then(() => {
                         // ok
-                    }, (p_reason) => {
+                    }, (_p_reason) => {
                         // Reveal failed. Retry refreshOutline once
                         this._refreshOutline(true, RevealType.RevealSelect);
                     });
@@ -4321,7 +4501,7 @@ export class LeoIntegration {
 
         const w_disposables: vscode.Disposable[] = [];
 
-        const q_minibufferQuickPick: Promise<vscode.QuickPickItem | undefined> = new Promise((resolve, reject) => {
+        const q_minibufferQuickPick: Promise<vscode.QuickPickItem | undefined> = new Promise((resolve, _reject) => {
             const quickPick = vscode.window.createQuickPick();
             quickPick.items = w_choices;
             quickPick.placeholder = Constants.USER_MESSAGES.MINIBUFFER_PROMPT;
@@ -4334,7 +4514,7 @@ export class LeoIntegration {
                         quickPick.hide();
                     }
                 }),
-                quickPick.onDidAccept(accepted => {
+                quickPick.onDidAccept(_accepted => {
                     if (/^\d+$/.test(quickPick.value)) {
                         // * Was an integer
                         this.gotoGlobalLine(Number(quickPick.value));
@@ -4457,7 +4637,7 @@ export class LeoIntegration {
 
         const hib = vscode.window.createInputBox();
         this._hibLastValue = undefined; // Prepare for 'cancel' as default.
-        const q_headlineInputBox = new Promise<string | undefined>((p_resolve, p_reject) => {
+        const q_headlineInputBox = new Promise<string | undefined>((p_resolve, _p_reject) => {
 
             hib.ignoreFocusOut = !!p_options.ignoreFocusOut;
             hib.value = p_options.value!;
@@ -4562,7 +4742,7 @@ export class LeoIntegration {
             // MAYBE DETACHED BODY CHANGED THAT CONTENT!
             this._leoFileSystem.setNewBodyUriTime(utils.strToLeoUri(p_node.gnx));
 
-            return this.showBody(false, this.config.treeKeepFocus).catch((p_error) => {
+            return this.showBody(false, this.config.treeKeepFocus).catch((_p_error) => {
                 return Promise.resolve(); // intercept cancellation as success: next one is going to replace anyways.
             });
             // Voluntary exit
@@ -4735,7 +4915,6 @@ export class LeoIntegration {
      * @returns Promise back from command's execution, if added on stack, undefined otherwise.
      * (see command stack 'rules' in commandStack.ts)
      */
-
     public async nodeCommand(p_userCommand: UserCommand, p_isNavigation?: boolean): Promise<LeoBridgePackage | undefined> {
         // No forced vscode save-triggers for direct calls from extension.js
         await this.triggerBodySave();
@@ -5144,7 +5323,7 @@ export class LeoIntegration {
                 return this.sendAction(
                     Constants.LEOBRIDGE.SET_UA_MEMBER,
                     { name: w_name, value: w_uaVal }
-                ).then((p_resultTag: LeoBridgePackage) => {
+                ).then((_p_resultTag: LeoBridgePackage) => {
                     this.setupRefresh(
                         Focus.NoChange,
                         {
@@ -5468,7 +5647,7 @@ export class LeoIntegration {
             w_panelID = Constants.FIND_ID;
             w_panel = this._findPanelWebviewView;
         }
-        return vscode.commands.executeCommand(w_panelID + '.focus', p_options).then((p_result) => {
+        return vscode.commands.executeCommand(w_panelID + '.focus', p_options).then((_p_result) => {
             if (w_panel && w_panel.show && !w_panel.visible) {
                 w_panel.show(false);
             }
@@ -5670,7 +5849,7 @@ export class LeoIntegration {
     /**
      * Check if search input should be forced-focused again
      */
-    public checkForceFindFocus(p_fromInit: boolean): void {
+    public checkForceFindFocus(_p_fromInit: boolean): void {
         if (this._findNeedsFocus) {
             let message = 'selectFind';
             if (this._findNeedsFocus === 2) {
@@ -5912,7 +6091,7 @@ export class LeoIntegration {
         const w_startValue = this._lastSettingsUsed!.findText === Constants.USER_MESSAGES.FIND_PATTERN_HERE ? '' : this._lastSettingsUsed!.findText;
 
         try {
-            return await new Promise<unknown>((resolve, reject) => {
+            return await new Promise<unknown>((resolve, _reject) => {
                 const input = vscode.window.createInputBox();
                 input.title = w_searchTitle;
                 input.value = w_startValue;
@@ -6020,7 +6199,7 @@ export class LeoIntegration {
         const w_startReplace = this._lastSettingsUsed?.replaceText;
 
         return this._isBusyTriggerSave(false, true)
-            .then((p_saveResult) => {
+            .then((_p_saveResult) => {
                 return this._inputFindPattern(false, w_startValue)
                     .then((p_findString) => {
                         if (!p_findString) {
@@ -6169,7 +6348,7 @@ export class LeoIntegration {
         }
         await vscode.commands.executeCommand(w_panelID + '.focus');
 
-        return new Promise((p_resolve, p_reject) => {
+        return new Promise((p_resolve, _p_reject) => {
             setTimeout(() => {
                 if (this._findPanelWebviewView && this._findPanelWebviewView.visible) {
                     w_panel = this._findPanelWebviewView;
@@ -6309,8 +6488,8 @@ export class LeoIntegration {
                     if (p_inputResult) {
                         const w_line = parseInt(p_inputResult);
                         if (!isNaN(w_line)) {
-                            this.sendAction(Constants.LEOBRIDGE.GOTO_GLOBAL_LINE, { line: w_line }).then((p_resultGoto: LeoBridgePackage) => {
-                                this.setupRefresh(Focus.Body, { tree: true, body: true, states: true, });
+                            this.sendAction(Constants.LEOBRIDGE.GOTO_GLOBAL_LINE, { line: w_line }).then((_p_resultGoto: LeoBridgePackage) => {
+                                this.setupRefresh(Focus.Body, { tree: true, body: true, scroll: true, states: true, });
                                 this.launchRefresh();
                             });
                         }
@@ -6318,7 +6497,7 @@ export class LeoIntegration {
                 });
         } else if (!isNaN(p_lineNumber)) {
             // p_lineNumber is a number
-            this.sendAction(Constants.LEOBRIDGE.GOTO_GLOBAL_LINE, { line: p_lineNumber }).then((p_resultGoto: LeoBridgePackage) => {
+            this.sendAction(Constants.LEOBRIDGE.GOTO_GLOBAL_LINE, { line: p_lineNumber }).then((_p_resultGoto: LeoBridgePackage) => {
                 this.setupRefresh(Focus.Body, { tree: true, body: true, states: true, });
                 this.launchRefresh();
             });
@@ -6348,7 +6527,7 @@ export class LeoIntegration {
                     this.sendAction(
                         Constants.LEOBRIDGE.TAG_CHILDREN,
                         { tag: p_inputResult }
-                    ).then((p_resultTag: LeoBridgePackage) => {
+                    ).then((_p_resultTag: LeoBridgePackage) => {
                         this.setupRefresh(
                             Focus.NoChange,
                             {
@@ -6395,7 +6574,7 @@ export class LeoIntegration {
                     this.sendAction(
                         Constants.LEOBRIDGE.TAG_NODE,
                         w_param
-                    ).then((p_resultTag: LeoBridgePackage) => {
+                    ).then((_p_resultTag: LeoBridgePackage) => {
 
                         this.setupRefresh(
                             Focus.NoChange,
@@ -6464,7 +6643,7 @@ export class LeoIntegration {
                         this.sendAction(
                             Constants.LEOBRIDGE.REMOVE_TAG,
                             w_param
-                        ).then((p_resultTag: LeoBridgePackage) => {
+                        ).then((_p_resultTag: LeoBridgePackage) => {
                             this.setupRefresh(
                                 Focus.NoChange,
                                 {
@@ -6497,7 +6676,7 @@ export class LeoIntegration {
                     this.sendAction(
                         Constants.LEOBRIDGE.REMOVE_TAGS,
                         p_node ? utils.buildNodeCommand(p_node) : undefined
-                    ).then((p_resultTag: LeoBridgePackage) => {
+                    ).then((_p_resultTag: LeoBridgePackage) => {
                         this.setupRefresh(
                             Focus.NoChange,
                             {
@@ -6586,7 +6765,7 @@ export class LeoIntegration {
             if (w_trigger && !this._documentPaneReveal) {
                 this._documentPaneReveal = w_docView.reveal(p_documentNode, { select: true, focus: false })
                     .then(
-                        (p_result) => {
+                        (_p_result) => {
                             // Shown document node
                             this._documentPaneReveal = undefined;
                         },
@@ -6634,7 +6813,7 @@ export class LeoIntegration {
      */
     public saveAsLeoFile(p_fromOutline?: boolean): Promise<LeoBridgePackage | undefined> {
         return this._isBusyTriggerSave(true, true)
-            .then((p_saveResult) => {
+            .then((_p_saveResult) => {
                 if (this.leoStates.fileOpenedReady && this.lastSelectedNode) {
                     return this._leoFilesBrowser.getLeoFileUrl(true);
                 } else {
@@ -6690,7 +6869,7 @@ export class LeoIntegration {
      */
     public saveAsLeoJsFile(p_fromOutline?: boolean): Promise<LeoBridgePackage | undefined> {
         return this._isBusyTriggerSave(true, true)
-            .then((p_saveResult) => {
+            .then((_p_saveResult) => {
                 if (this.leoStates.fileOpenedReady && this.lastSelectedNode) {
                     return this._leoFilesBrowser.getLeoJsFileUrl();
                 } else {
@@ -6911,7 +7090,7 @@ export class LeoIntegration {
     public closeLeoFile(): Promise<boolean> {
         let w_removeLastFileName: string = '';
         return this._isBusyTriggerSave(true, true)
-            .then((p_saveResult) => {
+            .then((_p_saveResult) => {
                 if (!this.leoStates.fileOpenedReady) {
                     vscode.window.showInformationMessage(Constants.USER_MESSAGES.CLOSE_ERROR);
                     return Promise.reject(Constants.USER_MESSAGES.CLOSE_ERROR);
@@ -6987,7 +7166,7 @@ export class LeoIntegration {
                                         });
                                 }
                                 return w_savePromise.then(
-                                    (p_packageAfterSave) => {
+                                    (_p_packageAfterSave) => {
                                         return this.sendAction(
                                             Constants.LEOBRIDGE.CLOSE_FILE,
                                             { forced: true }
@@ -7098,7 +7277,7 @@ export class LeoIntegration {
      */
     public openLeoFile(p_leoFileUri?: vscode.Uri): Promise<LeoBridgePackage | undefined> {
         return this._isBusyTriggerSave(true, true)
-            .then((p_saveResult) => {
+            .then((_p_saveResult) => {
                 if (p_leoFileUri && p_leoFileUri.scheme && p_leoFileUri.scheme.startsWith("leo")) {
                     p_leoFileUri = undefined; // Was used in the editor/title menu!
                 }
@@ -7188,7 +7367,7 @@ export class LeoIntegration {
      */
     public importAnyFile(p_leoFileUri?: vscode.Uri): Thenable<unknown> {
         return this._isBusyTriggerSave(true, true)
-            .then((p_saveResult) => {
+            .then((_p_saveResult) => {
                 let q_importFile: Promise<LeoBridgePackage | undefined>; // Promise for opening a file
                 if (p_leoFileUri && p_leoFileUri.fsPath.trim()) {
                     const w_fixedFilePath: string = p_leoFileUri.fsPath.replace(/\\/g, '/');
@@ -7247,9 +7426,9 @@ export class LeoIntegration {
      * * Export Outline
      * Export all headlines to an external file.
      */
-    public exportHeadlines(p_exportFileUri?: vscode.Uri): Thenable<unknown> {
+    public exportHeadlines(_p_exportFileUri?: vscode.Uri): Thenable<unknown> {
         return this._isBusyTriggerSave(true, true)
-            .then((p_saveResult) => {
+            .then((_p_saveResult) => {
                 if (this.leoStates.fileOpenedReady && this.lastSelectedNode) {
                     return this._leoFilesBrowser.getExportFileUrl(
                         "Export Headlines",
@@ -7293,7 +7472,7 @@ export class LeoIntegration {
     public flattenOutline(): Thenable<unknown> {
 
         return this._isBusyTriggerSave(true, true)
-            .then((p_saveResult) => {
+            .then((_p_saveResult) => {
                 if (this.leoStates.fileOpenedReady && this.lastSelectedNode) {
                     return this._leoFilesBrowser.getExportFileUrl(
                         "Flatten Selected Outline",
@@ -7335,7 +7514,7 @@ export class LeoIntegration {
     public outlineToCweb(): Thenable<unknown> {
 
         return this._isBusyTriggerSave(true, true)
-            .then((p_saveResult) => {
+            .then((_p_saveResult) => {
                 if (this.leoStates.fileOpenedReady && this.lastSelectedNode) {
                     return this._leoFilesBrowser.getExportFileUrl(
                         "Outline To CWEB",
@@ -7378,7 +7557,7 @@ export class LeoIntegration {
     public outlineToNoweb(): Thenable<unknown> {
 
         return this._isBusyTriggerSave(true, true)
-            .then((p_saveResult) => {
+            .then((_p_saveResult) => {
                 if (this.leoStates.fileOpenedReady && this.lastSelectedNode) {
                     return this._leoFilesBrowser.getExportFileUrl(
                         "Outline To Noweb",
@@ -7422,7 +7601,7 @@ export class LeoIntegration {
         // Convert one or more files, replacing the original files while removing any sentinels they contain.
 
         return this._isBusyTriggerSave(true, true)
-            .then((p_saveResult) => {
+            .then((_p_saveResult) => {
                 let q_importFiles: Promise<LeoBridgePackage | undefined>; // Promise for opening a file
                 if (p_leoFileUri && p_leoFileUri.fsPath.trim()) {
                     const w_fixedFilePath: string = p_leoFileUri.fsPath.replace(/\\/g, '/');
@@ -7494,7 +7673,7 @@ export class LeoIntegration {
     public weave(): Thenable<unknown> {
 
         return this._isBusyTriggerSave(true, true)
-            .then((p_saveResult) => {
+            .then((_p_saveResult) => {
                 if (this.leoStates.fileOpenedReady && this.lastSelectedNode) {
                     return this._leoFilesBrowser.getExportFileUrl(
                         "Weave",
@@ -7556,7 +7735,7 @@ export class LeoIntegration {
             q_fileName = Promise.resolve(fileName);
         } else {
             q_fileName = this._isBusyTriggerSave(true, true)
-                .then((p_saveResult) => {
+                .then((_p_saveResult) => {
                     if (this.leoStates.fileOpenedReady && this.lastSelectedNode) {
                         return this._leoFilesBrowser.getExportFileUrl(
                             "Write file from node",
@@ -7604,7 +7783,7 @@ export class LeoIntegration {
     public readFileIntoNode(p_leoFileUri?: vscode.Uri): Thenable<unknown> {
 
         return this._isBusyTriggerSave(true, true)
-            .then((p_saveResult) => {
+            .then((_p_saveResult) => {
                 let q_importFile: Promise<LeoBridgePackage | undefined>; // Promise for opening a file
                 if (p_leoFileUri && p_leoFileUri.fsPath.trim()) {
                     const w_fixedFilePath: string = p_leoFileUri.fsPath.replace(/\\/g, '/');
@@ -7673,7 +7852,7 @@ export class LeoIntegration {
      */
     public clickAtButton(p_node: LeoButtonNode): Promise<boolean> {
         return this._isBusyTriggerSave(false)
-            .then((p_saveResult) => {
+            .then((_p_saveResult) => {
 
                 if (p_node.rclicks.length) {
                     // Has rclicks so show menu to choose
@@ -7711,7 +7890,7 @@ export class LeoIntegration {
                     );
                 }
             })
-            .then((p_clickButtonResult: LeoBridgePackage) => {
+            .then((_p_clickButtonResult: LeoBridgePackage) => {
                 return this.sendAction(Constants.LEOBRIDGE.DO_NOTHING);
             })
             .then((p_package) => {
@@ -7729,7 +7908,7 @@ export class LeoIntegration {
                 );
                 this.launchRefresh();
                 return Promise.resolve(true);
-            }, (p_error) => {
+            }, (_p_error) => {
                 // Pass to catch the rejection if escaped by the user
                 return false;
             });
@@ -7864,7 +8043,7 @@ export class LeoIntegration {
                 () => {
                     this._undoPaneReveal = undefined;
                 },
-                (p_error) => {
+                (_p_error) => {
                     this._undoPaneReveal = undefined;
                     console.log('setUndoSelection could not reveal');
                 }